rename slackbot -> slack_utils
